containers/create: use shareable ipc by default for older clients
Older clients are not aware of new IPC modes (`shareable`, `private`,
and `none`, introduced in commit 7120976d7419) and expect the
implicit default to be what is now called `shareable`. The new daemon
default is now `private`, which is great, except it locks out older
clients from using `--ipc container:NNN` (as it now requires a donor
container created with explicit `--ipc shareable` which is
impossible to do with an old client).

To summarize:

 * clients with API < 1.32 create 'shareable' containers by default,
   irrespective of the daemon's default-ipc-mode;

 * clients with API >= 1.32 create containers with daemon's
   default-ipc-mode, which is now `private` unless overriden
   by a daemon command line option or a `daemon.json` parameter.

Test cases are modified accordingly.

Signed-off-by: Kir Kolyshkin <kolyshkin@gmail.com>